removed optimization methods that are not actually minimizing
In ShapeFitter we doesn't compute derivatives of the merit functions, so
several methods cannot be used.
Testing the result value from minimize() to check if minimization
actually occured is not trivial! Checking success flag is not enough.
I added a basic check to raise an exception in case of error, but I'm
not at all sure that I'm covering every wrong cases.